Rename file according to classname and put the enum out of the class
Renamed the file and deleted one empty class from it.
Put the enum out of the class, so you can access it without instantiating the class.
Remark: The PriorityNum type should be changed to Priority (enum), but if you access the enum via binding in the xaml you would get a string.